fix: 리뷰 반영으로 일정·할일 전환 draft 로직 정리
