+++ src/features/Calendar/components/CustomCalendar/CalendarModals.tsx
@@ -3,11 +3,10 @@ import { useMemo, useState } from 'react'
 
 import { useDetailEventQuery } from '@/shared/hooks/query/useCalendarQueries'
 import type { CalendarEvent } from '@/shared/types/calendar/types'
-import type { RepeatConfigSchema } from '@/shared/types/event/event'
 import type { ItemEditorDraft } from '@/shared/types/modal/itemEditor'
-import { defaultRepeatConfig } from '@/shared/types/recurrence/repeat'
 import ScheduleEditorModal from '@/shared/ui/Modals/ScheduleEditor'
 import TodoEditorModal from '@/shared/ui/Modals/TodoEditor'
+import { buildDefaultItemEditorDraft } from '@/shared/utils'
 
 type CalendarModalsProps = {
   modalDate: string
@@ -27,39 +26,6 @@ type CalendarModalsProps = {
       allDay: boolean,
       occurrenceDate?: CalendarEvent['occurrenceDate'],
     ) => void
-  }
-}
-
-const pad2 = (value: number) => String(value).padStart(2, '0')
-
-const formatTimeFromDate = (value: Date) => `${pad2(value.getHours())}:${pad2(value.getMinutes())}`
-
-const getDefaultDraft = (
-  date: string,
-  initialType: 'todo' | 'schedule',
-  initialEvent?: CalendarEvent | null,
-): ItemEditorDraft => {
-  const baseStart = initialEvent?.start ? new Date(initialEvent.start) : new Date(date)
-  const baseEnd =
-    initialEvent?.end && new Date(initialEvent.end).getTime() !== baseStart.getTime()
-      ? new Date(initialEvent.end)
-      : new Date(baseStart.getTime() + 60 * 60 * 1000)
-
-  return {
-    title:
-      initialType === 'schedule' && initialEvent?.title === '새 일정'
-        ? ''
-        : (initialEvent?.title ?? ''),
-    description: initialEvent?.content ?? '',
-    startDate: baseStart,
-    endDate: baseEnd,
-    startTime: formatTimeFromDate(baseStart),
-    endTime: initialType === 'todo' ? formatTimeFromDate(baseStart) : formatTimeFromDate(baseEnd),
-    isAllday: initialEvent?.isAllDay ?? false,
-    eventColor: initialEvent?.color ?? (initialType === 'todo' ? 'GRAY' : 'BLUE'),
-    repeatConfig: defaultRepeatConfig as RepeatConfigSchema,
-    location: initialEvent?.location ?? '',
-    address: initialEvent?.address ?? null,
   }
 }
 
@@ -84,12 +50,12 @@ const DraftBackedModal = ({
   onCloseModal,
   eventActions,
 }: DraftBackedModalProps) => {
-  const isTodoModal = modalEvent?.type === 'todo'
   const activeEvent = detailEvent ?? modalEvent
+  const isTodoModal = activeEvent?.type === 'todo'
   const [draftValues, setDraftValues] = useState<ItemEditorDraft | null>(() =>
     isModalEditing
       ? null
-      : getDefaultDraft(modalDate, isTodoModal ? 'todo' : 'schedule', activeEvent),
+      : buildDefaultItemEditorDraft(modalDate, isTodoModal ? 'todo' : 'schedule', activeEvent),
   )
 
   if (isTodoModal) {
@@ -99,7 +65,7 @@ const DraftBackedModal = ({
         onClose={onCloseModal}
         mode={modalMode}
         eventId={modalEventId}
-        event={modalEvent}
+        event={activeEvent}
         showTypeTabs={!isModalEditing}
         draftValues={draftValues}
         onDraftChange={setDraftValues}
@@ -118,7 +84,7 @@ const DraftBackedModal = ({
       onClose={onCloseModal}
       mode={modalMode}
       eventId={modalEventId}
-      event={detailEvent ?? modalEvent}
+      event={activeEvent}
       isEditing={isModalEditing}
       showTypeTabs={!isModalEditing}
       draftValues={draftValues}
